Project Nearing Completion
Today, I applied in the first bit of HTML for the main page of the website, then worked on code for the second part, however, it doesn't seem as if I applied it correctly, so that will be done soon enough. Speaking of which, I feel like getting the rest of this project done over the weekend.